feat: add kuai-ts support
for HSL Picker and Hue Slider only, clearly waste way to much time on
this.

Made me fr revaluate my life plan because wtf was that.
Stay blocked on async and VNode -> DOM lifecycle for waaaay too long.
Finally I am quite happy on how it turn out, this is not the most clean
and seemless code to declare ui but hey, I did it myself.

It quite bizarre to hate the lib developper when the lib developper is
you.

F* kuai tho.